pyproject.toml
[build-system]
requires = ["hatchling>=1.8.0", "hatch-vcs>=0.3.0"]
build-backend = "hatchling.build"

[project]
name = "tiny_uie"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sonesuke", email = "iamsonesuke@gmail.com" }
]
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
]
license = { text = "MIT" }
keywords = []
dependencies = [
    "accelerate>=0.33",
    "datasets>=2.21",
    "numpy>=2.3.3",
    "protobuf>=6.32.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.2",
    "transformers>=4.43",
]

[project.urls]
"Homepage" = "https://github.com/sonesuke/tiny-uie"
"Bug Reports" = "https://github.com/sonesuke/tiny-uie/issues"
"Source" = "https://github.com/sonesuke/tiny-uie"

[tool.hatch.version]
source = "vcs"
raw-options = { local_scheme = "no-local-version", fallback_version = "0.1.0" }

[tool.hatch.build.targets.wheel]
packages = ["src/tiny_uie"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.10",
    "ty>=0.0.1a21",
]

[project.scripts]
train-span = "tiny_uie.train_span:main"
train-type = "tiny_uie.train_type:main"
train-rel = "tiny_uie.train_rel:main"
infer = "tiny_uie.infer:main"
infer-rel = "tiny_uie.infer_rel:main"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--doctest-modules",
]

[tool.coverage.run]
omit = []

[tool.ruff]
exclude = [".venv", "tests"]
target-version = "py313"
include = ["*.py"]
line-length = 120

[tool.ruff.lint]
select = [
  "C9",   # McCabe complexity
  "ANN",  # Type annotations
  "S",    # Security
  "E",    # Syntax errors
  "F",    # Pyflakes
  "W",    # Warnings
  "I",    # isort
  "D",    # Docstrings
  "N",    # Naming conventions
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "SIM",  # flake8-simplify
  "ARG",  # flake8-unused-arguments
  "RET",  # flake8-return
  "TRY",  # tryceratops
  "PTH",  # Use pathlib
  "PL",   # Pylint
]
ignore = [
  "D203",  # one-blank-line-before-class
  "D213",  # multi-line-summary-second-line
  "D407",  # dashed-underline-after-section
  "D408",  # section-underline-after-name
  "D409",  # section-underline-matches-section-length
  "S101",  # Use of assert (common in tests)
  "PLR0913", # Too many arguments
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["tiny_uie"]
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ty.environment]
python-version = "3.13"
